Unify utility classes in api and billing
