+++ api/src/main/java/com/epam/pipeline/manager/billing/InstanceBillingLoader.java
@@ -38,6 +38,7 @@ import java.util.stream.Stream;
 public class InstanceBillingLoader implements BillingLoader<InstanceBilling> {
 
     private final BillingHelper billingHelper;
+    private final BillingSecurityHelper billingSecurityHelper;
     private final BillingIndexHelper billingIndexHelper;
     private final PreferenceManager preferenceManager;
 
@@ -46,7 +47,7 @@ public class InstanceBillingLoader implements BillingLoader<InstanceBilling> {
                                             final BillingExportRequest request) {
         final LocalDate from = request.getFrom();
         final LocalDate to = request.getTo();
-        final Map<String, List<String>> filters = billingHelper.getFilters(request.getFilters());
+        final Map<String, List<String>> filters = billingSecurityHelper.getFilters(request.getFilters());
         final BillingDiscount discount = Optional.ofNullable(request.getDiscount()).orElseGet(BillingDiscount::empty);
         return billings(client, from, to, filters, discount, getPageSize());
     }
