+++ api/src/main/java/com/epam/pipeline/manager/billing/RunBillingLoader.java
@@ -37,6 +37,7 @@ import java.util.stream.Stream;
 public class RunBillingLoader implements BillingLoader<RunBilling> {
 
     private final BillingHelper billingHelper;
+    private final BillingSecurityHelper billingSecurityHelper;
     private final BillingIndexHelper billingIndexHelper;
     private final PreferenceManager preferenceManager;
 
@@ -45,7 +46,7 @@ public class RunBillingLoader implements BillingLoader<RunBilling> {
                                        final BillingExportRequest request) {
         final LocalDate from = request.getFrom();
         final LocalDate to = request.getTo();
-        final Map<String, List<String>> filters = billingHelper.getFilters(request.getFilters());
+        final Map<String, List<String>> filters = billingSecurityHelper.getFilters(request.getFilters());
         final BillingDiscount discount = Optional.ofNullable(request.getDiscount()).orElseGet(BillingDiscount::empty);
         return billings(client, from, to, filters, discount, getPageSize());
     }
