+++ api/src/main/java/com/epam/pipeline/manager/billing/StorageBillingLoader.java
@@ -41,6 +41,7 @@ import java.util.stream.Stream;
 public class StorageBillingLoader implements BillingLoader<StorageBilling> {
 
     private final BillingHelper billingHelper;
+    private final BillingSecurityHelper billingSecurityHelper;
     private final BillingIndexHelper billingIndexHelper;
     private final PreferenceManager preferenceManager;
     private final StorageBillingDetailsLoader storageBillingDetailsLoader;
@@ -50,7 +51,7 @@ public class StorageBillingLoader implements BillingLoader<StorageBilling> {
                                            final BillingExportRequest request) {
         final LocalDate from = request.getFrom();
         final LocalDate to = request.getTo();
-        final Map<String, List<String>> filters = billingHelper.getFilters(request.getFilters());
+        final Map<String, List<String>> filters = billingSecurityHelper.getFilters(request.getFilters());
         final BillingDiscount discount = Optional.ofNullable(request.getDiscount()).orElseGet(BillingDiscount::empty);
         return billings(client, from, to, filters, discount, getPageSize());
     }
